rename to byteorder
